Basic admin panel working
Folders are very messy at the moment. Lots of stuff commented out,
especially where the ember framework was located. Might still consider
leaving ember in case that I can use the in the game pages. Also need
to do a good amount of work on the routes and adding attributes for the
users.