error fix on admin panel
Since the fields first_name and last_name are not being used in the
form for creating/editing an admin, these fields were causing an error
when trying to create an admin from admin panel. Commenting it out
fixes the problem, since they are not being used. If these fields need
to be adapted back to the app, make sure to add fields to form and
comment lines out